faketreeeval: move more code from sqlbase
Release note: None
