+++ pkg/sql/faketreeeval/evalctx.go
@@ -8,7 +8,8 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package sqlbase
+// Package faketreeeval provides fake implementations of tree eval interfaces.
+package faketreeeval
 
 import (
 	"context"
